Migrate core architecture to RDDs
This migrates the core project model, persistence and facets.
Operations and clustering remain to be done.

+++ or-model/src/main/java/org/openrefine/model/Record.java
@@ -34,14 +34,14 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 package org.openrefine.model;
 
 public class Record {
-    final public int fromRowIndex;
-    final public int toRowIndex;
-    final public int recordIndex;
+    final public long fromRowIndex;
+    final public long toRowIndex;
+    final public long recordIndex;
 
     public Record(
-            int fromRowIndex,
-            int toRowIndex,
-            int recordIndex
+            long fromRowIndex,
+            long toRowIndex,
+            long recordIndex
     ) {
         this.fromRowIndex = fromRowIndex;
         this.toRowIndex = toRowIndex;
